BMI GPIO Fixes..

+++ include/linux/bmi/omap_bmi.h
@@ -1,7 +1,7 @@
 /* platform data for OMAP BMI slots */
 
 struct omap_bmi_platform_data {
-  short gpios[4];
+  u8 gpios[4];
   char spi_cs;
   char i2c_bus_no;
 };
